+++ src/CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 2.2.2
 * New filters: Multiple hits, Revive
+* Improve verbiage for Haste, Revive and Counter effects
+* Fixes for PSY Nami GP condition and burst
 
 ### 2.2.1
 * Grand Party team builder.